refactor(tour): consolidate Hamiltonian path into alg.tour with shared base
Address maintainer feedback on the dev-list thread (move 1 of 1, DRY).

Per the maintainer's preference, fold the new Hamiltonian path classes
into the existing alg.tour package next to the cycle/tour code rather
than introducing a parallel alg.hamiltonian package. The interface
itself remains alongside HamiltonianCycleAlgorithm in alg.interfaces.
File moves use git rename so history is preserved. The JMH classes
move from perf.hamiltonian to perf.tour and the pom and module-info
follow.

Apply the DRY review pattern from PR #1340 (shared verification helper
in BaseManyToManyShortestPathsTest) to the production side as well.
Introduce HamiltonianPathAlgorithmBase, mirroring the existing
HamiltonianCycleAlgorithmBase in shape, with the helpers shared by
all three solvers:

- requireNotEmpty(graph)
- singletonPath(graph)
- vertexListToPath(vertices, graph)
- buildAdjacency(graph, indexToVertex, vertexToIndex, directed)
  (collapses parallel edges and skips self-loops)

BacktrackingHamiltonianPath, HeldKarpHamiltonianPath, and
DagHamiltonianPath now extend the base instead of duplicating these
helpers; net deletion in the impl classes. Package documentation in
alg.tour now explicitly distinguishes the cycle/tour and path solver
families.

All 58 hamiltonian unit tests, the random oracle suite, and the full
jgrapht-core FastTestSuite (5253 tests) pass after the refactor; the
state-count diagnostic shows identical state counts per family, so
the algorithmic behaviour is unchanged.
