+++ jgrapht-core/src/main/java/org/jgrapht/alg/tour/HamiltonianPathSearchResult.java
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR LGPL-2.1-or-later
  */
-package org.jgrapht.alg.hamiltonian;
+package org.jgrapht.alg.tour;
 
 import org.jgrapht.*;
